requirements.txt
streamlit

langchain
langchain-community
langchain-text-splitters
langchain-groq

chromadb
sentence-transformers
pypdf

python-dotenv